refactor: organize command handlers into commands/ subdirectory
Move command handlers to dedicated commands/ subdirectory for better organization:
- handlers/start.py → handlers/commands/start.py
- handlers/preferences.py → handlers/commands/preferences.py

Update import paths in:
- handlers/__init__.py: Import from new command paths
- tests/test_handlers/test_*.py: Update all import and patch references

All tests pass with 87.54% coverage maintained.
