+++ src/areyouok_telegram/handlers/__init__.py
@@ -1,11 +1,11 @@
+from areyouok_telegram.handlers.commands.preferences import on_preferences_command
+from areyouok_telegram.handlers.commands.start import on_start_command
 from areyouok_telegram.handlers.errors import on_error_event
 from areyouok_telegram.handlers.globals import on_dynamic_response_callback
 from areyouok_telegram.handlers.globals import on_new_update
 from areyouok_telegram.handlers.messages import on_edit_message
 from areyouok_telegram.handlers.messages import on_message_react
 from areyouok_telegram.handlers.messages import on_new_message
-from areyouok_telegram.handlers.preferences import on_preferences_command
-from areyouok_telegram.handlers.start import on_start_command
 
 __all__ = [
     "on_new_update",
